touch-core: add wpasupplicant
As part of importing network-manager from Jammy [1], we inherited
demotion of wpasupplicant from 'Depends' to 'Recommends' [2]. As we
don't install recommended packages, this package becomes not installed.

Since it's very unlikely that an Android phone doesn't come with a
Wi-Fi, add back this package via the seed.

Reported by Jami Kettunen on the Telegram channel.

[1] https://gitlab.com/ubports/development/core/packaging/network-manager/-/merge_requests/7
[2] https://git.launchpad.net/network-manager/commit/debian/control?h=ubuntu/jammy&id=dd9f1b1cce9f090e7d90df6708e94e1f2d43cb5f