android9plus & mainline: Ship respective xwayland variant
Multiple usecases (incl. Libertine, Box64 and Wine) require the presence of an xwayland server.
Since we now have an xwayland implementation for hybris-based devices we can ship both mainline
and android9plus images with xwayland support enabled.

Note that xwayland-hybris has only been tested on android9, hence we're not going to ship it in
images for older Halium generations yet.