Move libqzxing3 from touch-core to sdk-libs
libqzxing-dev is already in sdk-libs-dev, so libqzxing3 should also be
in sdk-libs. This makes it clear that libqzxing is part of public API.

Yes, in practice this is added for Camera app. But the cat has been let
out of the bag, so the next best thing is to make it right.