sdk-libs-dev: replaces libboost1.58-dev with non-versioned libboost-dev
While Boost is not ABI-stable, it doesn't matter when non of the built
part is linked. Including only libboost-dev doesn't allow apps to link
against built part as it doesn't contain any of .so, so we should be
safe.

Also make it clear that only the header-only part of Boost are part of
SDK.