touch-{hybris,android9plus}: revert XWayland to default variant for now
It's been discovered that it's not simple to backport xwayland-hybris
patchset to the current version of XWayland, as XWayland has removed
support for multiple GLAMOR backends (as they no longer need it for
Nvidia's EGLStream support). After a discussion with Alfred, he
suggested that that we "go with regular Xwayland right now through shm.
The -hybris patchset was tested poorly".

Bug: https://gitlab.com/ubports/ubuntu-touch/-/issues/2216